requirements.txt
streamlit==1.10.0
tensorflow==2.8.0
Pillow==9.1.0
numpy==1.22.3
